refactor: Text renderes moved to Renderers folder
