+++ examples/code-only/Example10_StrideUI_DragAndDrop_BulletPhysics/Program.cs
@@ -1,6 +1,7 @@
 using Example10_StrideUI_DragAndDrop;
 using Stride.CommunityToolkit.Bullet;
 using Stride.CommunityToolkit.Engine;
+using Stride.CommunityToolkit.Renderers;
 using Stride.CommunityToolkit.Rendering.ProceduralModels;
 using Stride.CommunityToolkit.Skyboxes;
 using Stride.Core.Mathematics;
