+++ examples/code-only/Example_CubicleCalamity_BulletPhysics/CubeStacker.cs
@@ -5,6 +5,7 @@ using Example_CubicleCalamity_BulletPhysics;
 using Stride.CommunityToolkit.Bullet;
 using Stride.CommunityToolkit.Engine;
 using Stride.CommunityToolkit.Games;
+using Stride.CommunityToolkit.Renderers;
 using Stride.CommunityToolkit.Rendering.Compositing;
 using Stride.CommunityToolkit.Rendering.ProceduralModels;
 using Stride.Core.Mathematics;
